Restrict permissions to user
While in the manifest we specify 0644 as the permissions for
`/etc/znc/config/znc.conf`, the ZNC daemon automatically resets them to 0600.
This causes that on every `puppet apply` the file mode is changed to the
specified value causing the ZNC server to be restarted and the file mode to be
reset again to what ZNC wants.

This commit sets the default permissions for every config file to 0600, such
that they remain readable only by the ZNC server.